ECL-496: Improved currency input validation (#216)
* Use double apostrophe in messages.en

* ECL-496: Improved currency input validation

* ECL-496: Decision made for when to remove all whitespace in binding

* ECL-496: Fixed decimal regex and only allow £ at the beginning of input

* ECL-496: Convert ukRevenue to a double

* ECL-496: Updated ukRevenue hint in welsh

* ECL-496: Converted revenue to BigDecimal

* ECL-496: Updated content for welsh

+++ app/uk/gov/hmrc/economiccrimelevyregistration/models/Registration.scala
@@ -28,7 +28,7 @@ final case class Registration(
   amlSupervisor: Option[AmlSupervisor],
   relevantAp12Months: Option[Boolean],
   relevantApLength: Option[Int],
-  relevantApRevenue: Option[Long],
+  relevantApRevenue: Option[BigDecimal],
   revenueMeetsThreshold: Option[Boolean],
   incorporatedEntityJourneyData: Option[IncorporatedEntityJourneyData],
   soleTraderEntityJourneyData: Option[SoleTraderEntityJourneyData],
